+++ Cargo.toml
@@ -22,15 +22,16 @@ geolocation = ["dep:windows", "dep:futures-util", "dep:futures", "dep:web-sys", 
 color_scheme = ["dep:web-sys", "dep:wasm-bindgen", "dep:wasm-bindgen-futures"]
 utils = ["dep:async-broadcast", "dep:uuid"]
 i18n = ["dep:serde", "dep:serde_json", "dep:unic-langid"]
-storage = ["dep:serde", "dep:postcard", "dep:once_cell", "dep:web-sys", "web-sys/Document", "web-sys/Element", "web-sys/HtmlDocument", "web-sys/Storage", "dep:directories", "dep:yazi"]
+storage = ["dep:tokio", "dep:dioxus-signals", "dep:serde", "dep:postcard", "dep:once_cell", "dep:web-sys", "web-sys/Document", "web-sys/Element", "web-sys/HtmlDocument", "web-sys/Storage", "dep:directories", "dep:yazi"]
 all = ["clipboard", "notifications", "geolocation", "color_scheme", "utils", "i18n", "storage"]
+default = ["storage"]
 
 # CI testing
 wasm-testing = ["geolocation", "color_scheme", "utils", "i18n"]
 desktop-testing = ["clipboard", "notifications", "geolocation", "utils", "i18n"]
 
 [dependencies]
-dioxus = { version = "0.4" }
+dioxus = { git = "https://github.com/DioxusLabs/dioxus.git", rev = "1ba6ca39e3f4d4fcd3b4e75e3a739d8c35514e5b" }
 cfg-if = "1.0.0"
 
 # utils
@@ -51,7 +52,11 @@ unic-langid = { version = "0.9.1", features = ["serde"], optional = true }
 rustc-hash = "1.1.0"
 postcard = { version = "1.0.2", features = ["use-std"], optional = true }
 once_cell = { version = "1.17.0", optional = true }
+dioxus-signals = { git = "https://github.com/DioxusLabs/dioxus.git", rev = "1ba6ca39e3f4d4fcd3b4e75e3a739d8c35514e5b", features = [
+    "serialize",
+], optional = true }
 yazi = { version = "0.1.4", optional = true }
+tokio = { version = "1.32.0", features = ["sync"], optional = true }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 directories = { version = "4.0.1", optional = true }
@@ -60,7 +65,7 @@ directories = { version = "4.0.1", optional = true }
 windows = { version = "0.48.0", features = ["Foundation", "Devices_Geolocation"], optional = true }
 
 [target.'cfg(target_family = "wasm")'.dependencies]
-web-sys = { version = "0.3.60", features = ["Window", "MediaQueryList", "Navigator", "Geolocation", "PositionOptions"], optional = true }
+web-sys = { version = "0.3.60", features = ["Window", "MediaQueryList", "Navigator", "Geolocation", "PositionOptions", "EventTarget"], optional = true }
 wasm-bindgen = { version = "0.2.87", optional = true }
 wasm-bindgen-futures = { version = "0.4.35", optional = true}
 js-sys = "0.3.62"
