+++ src/storage/client_storage/mod.rs
@@ -1,0 +1,9 @@
+cfg_if::cfg_if! {
+    if #[cfg(target_family = "wasm")] {
+        pub mod web;
+        pub use web::*;
+    } else {
+        pub mod fs;
+        pub use fs::*;
+    }
+}
