+++ src/storage/mod.rs
@@ -1,20 +1,15 @@
 //! # dioxus-storage
-
 //! A library for handling local storage ergonomically in Dioxus
-
 //! ## Usage
-
 //! ```rust
 //! use dioxus_storage::use_storage;
 //! use dioxus::prelude::*;
-
 //! fn main() {
 //!     dioxus_web::launch(app)
 //! }
-
+//! 
 //! fn app(cx: Scope) -> Element {
 //!     let num = use_persistent(cx, "count", || 0);
-
 //!     cx.render(rsx! {
 //!         div {
 //!             button {
@@ -33,8 +28,10 @@
 
 mod client_storage;
 mod storage;
+mod persistence;
 
-pub use client_storage::{use_persistent, use_singleton_persistent};
+pub use persistence::{use_persistent, use_singleton_persistent};
+pub use client_storage::ClientStorage;
 
-#[cfg(not(target_arch = "wasm32"))]
+#[cfg(not(target_family = "wasm"))]
 pub use client_storage::set_dir;
